pyproject.toml
[project]
name = "plant-disease-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.4.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "pillow>=8.3.0",
    "scikit-learn>=0.24.0",
    "seaborn>=0.11.0",
    "tensorflow>=2.10.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
]
